Rename file for basic viewers.
